pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.7.0",
    "requests>=2.32.3",
    "trafilatura>=2.0.0",
]
